Extract core: monolith fully modularized, server.py now pure orchestration
Final extraction. The monolith is gone.

- book/_legacy.py renamed to book/core.py (git-tracked rename), class
  LegacyMixin renamed to CoreMixin. Pruned unused imports now that
  reconciliation, reporting, budgets, scheduling, investments, business,
  and admin have all moved out — core only needs date/timedelta,
  Decimal/InvalidOperation, piecash, and five base serializers.
- tools/core.py: register(mcp, get_book) attaches the 15 core tool
  wrappers (book summary, account CRUD, transaction CRUD + search +
  replace_splits). Same shape as every other module.
- book/__init__.py: _MIXIN_MAP now lists all eight modules including
  core. Dropped the LegacyMixin import and its hardcoded slot in
  build_book_class — no module is "always included at class-build time"
  anymore; every module goes through the same lazy-load path.
  _apply_module_filter still forces 'core' into the enabled set, so
  core is effectively always loaded at runtime — but through the same
  mechanism as any other module.
- server.py: 906 -> 516 lines. The 15 core tool wrappers are gone. The
  file is now pure orchestration: FastMCP setup, the instructions
  block, TOOL_MODULES, _validate_tool_modules, _lazy_load_tool_module,
  _apply_module_filter, get_book, main(), and the debug-only
  get_server_config tool + gnucash://accounts resource.
- tests/test_modules.py: replaced
  test_non_extracted_modules_are_fully_registered_at_import (trivially
  true now — the expected set is empty) with
  test_every_module_is_extracted, which asserts
  extracted_modules() == set(TOOL_MODULES.keys()).
- tests/test_tools.py: test_unexpected_error_handling was patching
  server_module.get_book to exercise safe_tool's catch-all, which worked
  when list_accounts lived at server.py module scope. After
  modularization list_accounts is a closure in tools/core.py that
  captured get_book at register time, so attribute patches on
  server_module can't reach it. Changed to patch GnuCashBook.list_accounts
  at the book layer — same test intent, same assertion.

Tools registered at import: 0. All 75 lazy-load via importlib when
their module is enabled. --modules is now fully truthful: disabled
modules never parse their Pydantic tool schemas into memory.

MRO: GnuCashBook -> AdminMixin -> BudgetsMixin -> BusinessMixin ->
CoreMixin -> InvestmentsMixin -> ReconciliationMixin -> ReportingMixin
-> SchedulingMixin -> BaseGnuCashBook.

Full suite: 602 passed. Pure code motion — no behavior change.
